refactor: extract shared handler utilities into common module
Extract shared helper functions from handlers/mod.rs into a new
handlers/common.rs module:

- Cache helpers: cache_read_dir, build_cache_location
- Transaction resolution: resolve_inputs_to_txs
- IDL pipeline: collect_program_ids, auto_fetch_missing_idls,
  run_idl_pipeline, prepare_accounts_and_idls
- Address validation: warn_unmatched_addresses and helpers
- All associated unit tests

Update decode.rs and simulate.rs to use the new common module paths.

Also add interactive help for account subcommand when called without
arguments (matching decode behavior).
